Organized frontend src directory
created components, context, auth, and styles directories to better organize frontend directory
