+++ teeup-frontend/src/components/auth/Signup.jsx
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { AuthContext } from './AuthContext';
+import { AuthContext } from '../../context/AuthContext';
 
 function Signup() {
     const [username, setUsername] = useState('');
