+++ teeup-frontend/src/components/common/NavBar.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { AuthContext } from './AuthContext';
+import { AuthContext } from '../../context/AuthContext';
 
 const NavBar = () => {
   const { auth, setAuth } = useContext(AuthContext);
