+++ teeup-frontend/src/components/common/WelcomeMessage.jsx
